LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/radius-expected.txt


PASS Can set 'r' to CSS-wide keywords
FAIL Can set 'r' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'r' to a percent Invalid values
FAIL Can set 'r' to a length Invalid values
PASS Setting 'r' to a time throws TypeError
PASS Setting 'r' to an angle throws TypeError
PASS Setting 'r' to a flexible length throws TypeError
FAIL Setting 'r' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'r' to a URL throws TypeError
PASS Setting 'r' to a transform throws TypeError
PASS Can set 'rx' to CSS-wide keywords
FAIL Can set 'rx' to var() references assert_equals: expected 2 but got 1
PASS Can set 'rx' to the 'auto' keyword
FAIL Can set 'rx' to a percent Invalid values
FAIL Can set 'rx' to a length Invalid values
PASS Setting 'rx' to a time throws TypeError
PASS Setting 'rx' to an angle throws TypeError
PASS Setting 'rx' to a flexible length throws TypeError
FAIL Setting 'rx' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'rx' to a URL throws TypeError
PASS Setting 'rx' to a transform throws TypeError
PASS Can set 'ry' to CSS-wide keywords
FAIL Can set 'ry' to var() references assert_equals: expected 2 but got 1
PASS Can set 'ry' to the 'auto' keyword
FAIL Can set 'ry' to a percent Invalid values
FAIL Can set 'ry' to a length Invalid values
PASS Setting 'ry' to a time throws TypeError
PASS Setting 'ry' to an angle throws TypeError
PASS Setting 'ry' to a flexible length throws TypeError
FAIL Setting 'ry' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'ry' to a URL throws TypeError
PASS Setting 'ry' to a transform throws TypeError
